btrfs-progs: btrfs-debug-tree: handle corruption more gracefully
This patch fixes the same thing in two different places. First,
the first of the two BUG() tests is just a special case of the
second one and can therefore be omitted. Second, instead of bailing
out with BUG(), just print a reasonable error message and check the
next child.

Signed-off-by: Martin Wilck <mwilck@arcor.de>
Signed-off-by: David Sterba <dsterba@suse.cz>